@@ -4,8 +4,8 @@ version: 0.2.0
 homepage: https://fair.58.com/
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  test: ^1.6.0
+  pedantic: ^1.11.1
+  test: ^1.17.12
