@@ -4,31 +4,31 @@ version: 0.3.0+1
 homepage: https://fair.58.com/
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 
 #executables:
 #  fairc:
 dependencies:
-  source_gen: ^0.9.4+5
-  build: ^1.3.0
-  crypto: ^2.0.0
-  build_config: '>=0.2.6 <0.5.0'
-  process: ^3.0.13
-  archive: ^2.0.0
-  intl: ^0.16.1
-  path: ^1.6.4
-  analyzer: ^0.41.2
-  dart_style: ^1.3.6
-  glob: ^1.2.0
-  pedantic: ^1.8.0
+  source_gen: ^1.0.0
+  build: ^2.1.0
+  crypto: ^3.0.1
+  build_config: ^1.0.0
+  process: ^4.0.0
+  archive: ^3.3.0
+  intl: ^0.17.0
+  path: ^1.8.0
+  analyzer: ^1.5.0
+  dart_style: ^2.0.0
+  glob: ^2.0.1
+  pedantic: ^1.11.1
 #  fair_annotation: ^0.2.0
   fair_annotation:
     path: ../annotation
-  http: ^0.12.2
+  http: ^0.13.3
   fairc:
     path: ../dart2dsl
   dartToJs:
     path: ../dart2js
 
 dev_dependencies:
-  test: ^1.6.0
+  test: ^1.17.12
