@@ -4,15 +4,14 @@ description: A sample command-line application.
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0 <3.0.0"
 executables:
   fairc:
 dependencies:
-  analyzer: ^0.41.2
-
+  analyzer: ^1.5.0
 
 dev_dependencies:
-  pedantic: ^1.8.0
-  test: ^1.6.0
-  source_gen: ^0.9.7+1
-  build_runner: ^1.4.0
+  pedantic: ^1.11.1
+  test: ^1.17.12
+  source_gen: ^1.0.0
+  build_runner: ^2.0.0
