@@ -4,8 +4,7 @@ version: 2.0.0
 homepage: https://fair.58.com/
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -17,17 +16,17 @@ dependencies:
 
 #  fair_version: ^0.2.0
   fair_version:
-    path: ../flutter_version/flutter_1_22_4
+    path: ../flutter_version/flutter_2_0_6
 
-  flat_buffers: ^1.12.0
-  url_launcher: ^5.7.2
-  http: ^0.12.2
+  flat_buffers: ^2.0.5
+  url_launcher: ^6.0.10
+  http: ^0.13.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  analyzer: ^0.38.5
-  pedantic: ^1.9.2
+  analyzer: ^1.5.0
+  pedantic: ^1.11.1
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
