@@ -17,13 +17,12 @@ class Reference {
   final ByteData _buffer;
   final int _offset;
   final BitWidth _parentWidth;
-  final String _path;
-  int _byteWidth;
-  ValueType _valueType;
-  int _length;
+  final String? _path;
+  late int _byteWidth;
+  late ValueType _valueType;
+  int? _length;
 
-  Reference._(this._buffer, this._offset, this._parentWidth, int packedType,
-      this._path) {
+  Reference._(this._buffer, this._offset, this._parentWidth, int packedType, this._path) {
     _byteWidth = 1 << (packedType & 3);
     _valueType = ValueTypeUtils.fromInt(packedType >> 2);
   }
@@ -38,28 +37,23 @@ class Reference {
     final byteWidth = byteData.getUint8(len - 1);
     final packedType = byteData.getUint8(len - 2);
     final offset = len - byteWidth - 2;
-    return Reference._(ByteData.view(buffer), offset,
-        BitWidthUtil.fromByteWidth(byteWidth), packedType, '/');
+    return Reference._(ByteData.view(buffer), offset, BitWidthUtil.fromByteWidth(byteWidth), packedType, '/');
   }
 
   /// Returns true if the underlying value is null.
   bool get isNull => _valueType == ValueType.Null;
 
   /// Returns true if the underlying value can be represented as [num].
-  bool get isNum =>
-      ValueTypeUtils.isNumber(_valueType) ||
-      ValueTypeUtils.isIndirectNumber(_valueType);
+  bool get isNum => ValueTypeUtils.isNumber(_valueType) || ValueTypeUtils.isIndirectNumber(_valueType);
 
   /// Returns true if the underlying value was encoded as a float (direct or indirect).
-  bool get isDouble =>
-      _valueType == ValueType.Float || _valueType == ValueType.IndirectFloat;
+  bool get isDouble => _valueType == ValueType.Float || _valueType == ValueType.IndirectFloat;
 
   /// Returns true if the underlying value was encoded as an int or uint (direct or indirect).
   bool get isInt => isNum && !isDouble;
 
   /// Returns true if the underlying value was encoded as a string or a key.
-  bool get isString =>
-      _valueType == ValueType.String || _valueType == ValueType.Key;
+  bool get isString => _valueType == ValueType.String || _valueType == ValueType.Key;
 
   /// Returns true if the underlying value was encoded as a bool.
   bool get isBool => _valueType == ValueType.Bool;
@@ -74,7 +68,7 @@ class Reference {
   bool get isMap => _valueType == ValueType.Map;
 
   /// If this [isBool], returns the bool value. Otherwise, returns null.
-  bool get boolValue {
+  bool? get boolValue {
     if (_valueType == ValueType.Bool) {
       return _readInt(_offset, _parentWidth) != 0;
     }
@@ -84,7 +78,7 @@ class Reference {
   /// Returns an [int], if the underlying value can be represented as an int.
   ///
   /// Otherwise returns [null].
-  int get intValue {
+  int? get intValue {
     if (_valueType == ValueType.Int) {
       return _readInt(_offset, _parentWidth);
     }
@@ -103,7 +97,7 @@ class Reference {
   /// Returns [double], if the underlying value [isDouble].
   ///
   /// Otherwise returns [null].
-  double get doubleValue {
+  double? get doubleValue {
     if (_valueType == ValueType.Float) {
       return _readFloat(_offset, _parentWidth);
     }
@@ -116,12 +110,12 @@ class Reference {
   /// Returns [num], if the underlying value is numeric, be it int uint, or float (direct or indirect).
   ///
   /// Otherwise returns [null].
-  num get numValue => doubleValue ?? intValue;
+  num? get numValue => doubleValue ?? intValue;
 
   /// Returns [String] value or null otherwise.
   ///
   /// This method performers a utf8 decoding, as FlexBuffers format stores strings in utf8 encoding.
-  String get stringValue {
+  String? get stringValue {
     if (_valueType == ValueType.String || _valueType == ValueType.Key) {
       return utf8.decode(_buffer.buffer.asUint8List(_indirect, length));
     }
@@ -129,7 +123,7 @@ class Reference {
   }
 
   /// Returns [Uint8List] value or null otherwise.
-  Uint8List get blobValue {
+  Uint8List? get blobValue {
     if (_valueType == ValueType.Blob) {
       return _buffer.buffer.asUint8List(_indirect, length);
     }
@@ -144,27 +138,21 @@ class Reference {
     if (key is int && ValueTypeUtils.isAVector(_valueType)) {
       final index = key;
       if (index >= length || index < 0) {
-        throw ArgumentError(
-            'Key: [$key] is not applicable on: $_path of: $_valueType length: $length');
+        throw ArgumentError('Key: [$key] is not applicable on: $_path of: $_valueType length: $length');
       }
       final elementOffset = _indirect + index * _byteWidth;
-      final reference = Reference._(_buffer, elementOffset,
-          BitWidthUtil.fromByteWidth(_byteWidth), 0, '$_path[$index]');
+      final reference = Reference._(_buffer, elementOffset, BitWidthUtil.fromByteWidth(_byteWidth), 0, '$_path[$index]');
       reference._byteWidth = 1;
       if (ValueTypeUtils.isTypedVector(_valueType)) {
-        reference._valueType =
-            ValueTypeUtils.typedVectorElementType(_valueType);
+        reference._valueType = ValueTypeUtils.typedVectorElementType(_valueType);
         return reference;
       }
       if (ValueTypeUtils.isFixedTypedVector(_valueType)) {
-        reference._valueType =
-            ValueTypeUtils.fixedTypedVectorElementType(_valueType);
+        reference._valueType = ValueTypeUtils.fixedTypedVectorElementType(_valueType);
         return reference;
       }
-      final packedType =
-          _buffer.getUint8(_indirect + length * _byteWidth + index);
-      return Reference._(_buffer, elementOffset,
-          BitWidthUtil.fromByteWidth(_byteWidth), packedType, '$_path[$index]');
+      final packedType = _buffer.getUint8(_indirect + length * _byteWidth + index);
+      return Reference._(_buffer, elementOffset, BitWidthUtil.fromByteWidth(_byteWidth), packedType, '$_path[$index]');
     }
     if (key is String && _valueType == ValueType.Map) {
       final index = _keyIndex(key);
@@ -172,8 +160,7 @@ class Reference {
         return _valueForIndexWithKey(index, key);
       }
     }
-    throw ArgumentError(
-        'Key: [$key] is not applicable on: $_path of: $_valueType');
+    throw ArgumentError('Key: [$key] is not applicable on: $_path of: $_valueType');
   }
 
   /// Get an iterable if the underlying flexBuffer value is a vector.
@@ -210,27 +197,22 @@ class Reference {
   /// If the values is a string or a blob, the length reflects a number of bytes the value occupies (strings are encoded in utf8 format).
   int get length {
     if (_length != null) {
-      return _length;
+      return _length!;
     }
     // needs to be checked before more generic isAVector
     if (ValueTypeUtils.isFixedTypedVector(_valueType)) {
       _length = ValueTypeUtils.fixedTypedVectorElementSize(_valueType);
-    } else if (_valueType == ValueType.Blob ||
-        ValueTypeUtils.isAVector(_valueType) ||
-        _valueType == ValueType.Map) {
-      _length = _readUInt(
-          _indirect - _byteWidth, BitWidthUtil.fromByteWidth(_byteWidth));
+    } else if (_valueType == ValueType.Blob || ValueTypeUtils.isAVector(_valueType) || _valueType == ValueType.Map) {
+      _length = _readUInt(_indirect - _byteWidth, BitWidthUtil.fromByteWidth(_byteWidth));
     } else if (_valueType == ValueType.Null) {
       _length = 0;
     } else if (_valueType == ValueType.String) {
       final indirect = _indirect;
       var sizeByteWidth = _byteWidth;
-      var size = _readUInt(
-          indirect - sizeByteWidth, BitWidthUtil.fromByteWidth(sizeByteWidth));
+      var size = _readUInt(indirect - sizeByteWidth, BitWidthUtil.fromByteWidth(sizeByteWidth));
       while (_buffer.getInt8(indirect + size) != 0) {
         sizeByteWidth <<= 1;
-        size = _readUInt(indirect - sizeByteWidth,
-            BitWidthUtil.fromByteWidth(sizeByteWidth));
+        size = _readUInt(indirect - sizeByteWidth, BitWidthUtil.fromByteWidth(sizeByteWidth));
       }
       _length = size;
     } else if (_valueType == ValueType.Key) {
@@ -243,7 +225,7 @@ class Reference {
     } else {
       _length = 1;
     }
-    return _length;
+    return _length!;
   }
 
   /// Returns a minified JSON representation of the underlying FlexBuffer value.
@@ -253,7 +235,7 @@ class Reference {
   /// Blob values are represented as base64 encoded string.
   String get json {
     if (_valueType == ValueType.Bool) {
-      return boolValue ? 'true' : 'false';
+      return boolValue==true ? 'true' : 'false';
     }
     if (_valueType == ValueType.Null) {
       return 'null';
@@ -265,7 +247,7 @@ class Reference {
       return jsonEncode(stringValue);
     }
     if (_valueType == ValueType.Blob) {
-      return jsonEncode(base64Encode(blobValue));
+      return jsonEncode(base64Encode(blobValue??[]));
     }
     if (ValueTypeUtils.isAVector(_valueType)) {
       final result = StringBuffer();
@@ -293,8 +275,7 @@ class Reference {
       result.write('}');
       return result.toString();
     }
-    throw UnsupportedError(
-        'Type: $_valueType is not supported for JSON conversion');
+    throw UnsupportedError('Type: $_valueType is not supported for JSON conversion');
   }
 
   /// Computes the indirect offset of the value.
@@ -349,20 +330,16 @@ class Reference {
   }
 
   void _validateOffset(int offset, BitWidth width) {
-    if (_offset < 0 ||
-        _buffer.lengthInBytes <= offset + width.index ||
-        offset & (BitWidthUtil.toByteWidth(width) - 1) != 0) {
+    if (_offset < 0 || _buffer.lengthInBytes <= offset + width.index || offset & (BitWidthUtil.toByteWidth(width) - 1) != 0) {
       throw StateError('Bad offset: $offset, width: $width');
     }
   }
 
-  int _keyIndex(String key) {
+  int? _keyIndex(String key) {
     final input = utf8.encode(key);
     final keysVectorOffset = _indirect - _byteWidth * 3;
-    final indirectOffset = keysVectorOffset -
-        _readUInt(keysVectorOffset, BitWidthUtil.fromByteWidth(_byteWidth));
-    final byteWidth = _readUInt(
-        keysVectorOffset + _byteWidth, BitWidthUtil.fromByteWidth(_byteWidth));
+    final indirectOffset = keysVectorOffset - _readUInt(keysVectorOffset, BitWidthUtil.fromByteWidth(_byteWidth));
+    final byteWidth = _readUInt(keysVectorOffset + _byteWidth, BitWidthUtil.fromByteWidth(_byteWidth));
     var low = 0;
     var high = length - 1;
     while (low <= high) {
@@ -380,8 +357,7 @@ class Reference {
 
   int _diffKeys(List<int> input, int index, int indirectOffset, int byteWidth) {
     final keyOffset = indirectOffset + index * byteWidth;
-    final keyIndirectOffset =
-        keyOffset - _readUInt(keyOffset, BitWidthUtil.fromByteWidth(byteWidth));
+    final keyIndirectOffset = keyOffset - _readUInt(keyOffset, BitWidthUtil.fromByteWidth(byteWidth));
     for (var i = 0; i < input.length; i++) {
       final dif = input[i] - _buffer.getUint8(keyIndirectOffset + i);
       if (dif != 0) {
@@ -395,27 +371,22 @@ class Reference {
     final indirect = _indirect;
     final elementOffset = indirect + index * _byteWidth;
     final packedType = _buffer.getUint8(indirect + length * _byteWidth + index);
-    return Reference._(_buffer, elementOffset,
-        BitWidthUtil.fromByteWidth(_byteWidth), packedType, '$_path/$key');
+    return Reference._(_buffer, elementOffset, BitWidthUtil.fromByteWidth(_byteWidth), packedType, '$_path/$key');
   }
 
   Reference _valueForIndex(int index) {
     final indirect = _indirect;
     final elementOffset = indirect + index * _byteWidth;
     final packedType = _buffer.getUint8(indirect + length * _byteWidth + index);
-    return Reference._(_buffer, elementOffset,
-        BitWidthUtil.fromByteWidth(_byteWidth), packedType, '$_path/[$index]');
+    return Reference._(_buffer, elementOffset, BitWidthUtil.fromByteWidth(_byteWidth), packedType, '$_path/[$index]');
   }
 
   String _keyForIndex(int index) {
     final keysVectorOffset = _indirect - _byteWidth * 3;
-    final indirectOffset = keysVectorOffset -
-        _readUInt(keysVectorOffset, BitWidthUtil.fromByteWidth(_byteWidth));
-    final byteWidth = _readUInt(
-        keysVectorOffset + _byteWidth, BitWidthUtil.fromByteWidth(_byteWidth));
+    final indirectOffset = keysVectorOffset - _readUInt(keysVectorOffset, BitWidthUtil.fromByteWidth(_byteWidth));
+    final byteWidth = _readUInt(keysVectorOffset + _byteWidth, BitWidthUtil.fromByteWidth(_byteWidth));
     final keyOffset = indirectOffset + index * byteWidth;
-    final keyIndirectOffset =
-        keyOffset - _readUInt(keyOffset, BitWidthUtil.fromByteWidth(byteWidth));
+    final keyIndirectOffset = keyOffset - _readUInt(keyOffset, BitWidthUtil.fromByteWidth(byteWidth));
     var length = 0;
     while (_buffer.getUint8(keyIndirectOffset + length) != 0) {
       length += 1;
@@ -424,11 +395,9 @@ class Reference {
   }
 }
 
-class _VectorIterator
-    with IterableMixin<Reference>
-    implements Iterator<Reference> {
+class _VectorIterator with IterableMixin<Reference> implements Iterator<Reference> {
   final Reference _vector;
-  int index;
+  late int index;
 
   _VectorIterator(this._vector) {
     index = -1;
@@ -449,7 +418,7 @@ class _VectorIterator
 
 class _MapKeyIterator with IterableMixin<String> implements Iterator<String> {
   final Reference _map;
-  int index;
+  late int index;
 
   _MapKeyIterator(this._map) {
     index = -1;
@@ -468,11 +437,9 @@ class _MapKeyIterator with IterableMixin<String> implements Iterator<String> {
   Iterator<String> get iterator => this;
 }
 
-class _MapValueIterator
-    with IterableMixin<Reference>
-    implements Iterator<Reference> {
+class _MapValueIterator with IterableMixin<Reference> implements Iterator<Reference> {
   final Reference _map;
-  int index;
+  late int index;
 
   _MapValueIterator(this._map) {
     index = -1;
