@@ -14,7 +14,7 @@ import 'package:flutter/services.dart';
 import 'package:fair_version/fair_version.dart';
 
 typedef VoidMsgCallback = void Function();
-typedef StringMsgCallback = String Function(String msg);
+typedef StringMsgCallback = String? Function(String? msg);
 
 final DynamicLibrary dl = Platform.isAndroid
     ? DynamicLibrary.open('libfairflutter.so')
@@ -29,8 +29,8 @@ class FairMessageChannel {
           'invokeJSCommonFuncSync')
       .asFunction();
 
-  BasicMessageChannel<String> _commonChannel;
-  MethodChannel _methodChannel;
+  BasicMessageChannel<String?>? _commonChannel;
+  MethodChannel? _methodChannel;
 
   factory FairMessageChannel() {
     return _msg;
@@ -44,13 +44,13 @@ class FairMessageChannel {
 
   void _initMessageChannel() {
     _commonChannel ??=
-        BasicMessageChannel(COMMON_MESSAGE_CHANNEL, StringCodec());
+        BasicMessageChannel<String?>(COMMON_MESSAGE_CHANNEL, StringCodec());
     _methodChannel ??= MethodChannel(JS_LOADER);
 
-    _commonChannel.setMessageHandler((String message) async {
+    _commonChannel!.setMessageHandler((String? message) async {
       print('来自native端的消息：$message');
       //js 异步调用dart中的相关方法
-      var data = json.decode(message);
+      var data = json.decode(message??'');
       var funcName = data['funcName']?.toString();
 
       if (funcName == 'invokePlugin') {
@@ -62,25 +62,25 @@ class FairMessageChannel {
       return 'reply from dart';
     });
 
-    _methodChannel.setMethodCallHandler((call) async {});
+    _methodChannel!.setMethodCallHandler((call) async {});
   }
 
-  StringMsgCallback _callback;
+  StringMsgCallback? _callback;
 
   void setMessageHandler(StringMsgCallback callback) {
     _callback = callback;
   }
 
-  Future<dynamic> loadJS(String args, StringMsgCallback callback) {
-    return _methodChannel.invokeMethod('loadMainJs', args);
+  Future<dynamic> loadJS(String args, StringMsgCallback? callback) {
+    return _methodChannel!.invokeMethod('loadMainJs', args);
   }
 
-  Future<dynamic> release(String args, VoidMsgCallback callback) {
-    return _methodChannel.invokeMethod('releaseMainJs', args);
+  Future<dynamic> release(String args, VoidMsgCallback? callback) {
+    return _methodChannel!.invokeMethod('releaseMainJs', args);
   }
 
-  Future<String> sendCommonMessage(dynamic msg) async {
-    return _commonChannel.send(msg);
+  Future<String?> sendCommonMessage(dynamic msg) async {
+    return _commonChannel!.send(msg);
   }
 
   dynamic sendCommonMessageSync(dynamic msg) =>
