@@ -11,21 +11,30 @@ import 'dart:math';
 import 'package:archive/archive.dart';
 import 'package:build/build.dart';
 import 'package:crypto/crypto.dart' show md5;
+import 'package:fair_compiler/src/state_transfer.dart';
 import 'package:path/path.dart' as path;
 import 'package:dartToJs/index.dart' as dart2js;
-import 'helper.dart' show FlatCompiler;
+import 'helper.dart' show FlatCompiler, ModuleNameHelper;
 
 class ArchiveBuilder extends PostProcessBuilder with FlatCompiler {
   @override
   FutureOr<void> build(PostProcessBuildStep buildStep) async {
-    final dir = path.join('build', 'fair');
+    final dir = path.join('build','fair');
     Directory(dir).createSync(recursive: true);
+
+    /// Get the module value of the bundle file,
+    /// the module value will be used as the prefix of the bundle file name.
+    /// Example：'home_recommend.fair.json'
+    var moduleNameKey = buildStep.inputId.path.replaceAll('.bundle.json', '');
+    var moduleNameValue = ModuleNameHelper().modules[moduleNameKey];
+
     final bundleName = path.join(
-        dir,
-        buildStep.inputId.path
-            .replaceAll(inputExtensions.first, '.fair.json')
-            .replaceAll('/', '_')
-            .replaceAll('\\', '_'));
+            dir,
+            buildStep.inputId.path
+                .replaceAll(inputExtensions.first, '.fair.json')
+                .replaceAll('lib', moduleNameValue)
+                .replaceAll('/', '_')
+                .replaceAll('\\', '_'));
     final jsName = bundleName.replaceFirst('.json', '.js');
 
     await dart2JS(buildStep.inputId.path, jsName);
@@ -33,30 +42,29 @@ class ArchiveBuilder extends PostProcessBuilder with FlatCompiler {
 
     // 压缩下发产物
     var zipSrcPath = path.join(Directory.current.path, 'build', 'fair');
-    var zipDesPath =
-        path.join(Directory.current.path, 'build', 'fair', 'fair_patch.zip');
+    var zipDesPath = path.join(Directory.current.path, 'build', 'fair', 'fair_patch.zip');
     _zip(Directory(zipSrcPath), File(zipDesPath));
+
+   await stateTransfer();
   }
 
   @override
   Iterable<String> get inputExtensions => ['.bundle.json'];
 
-  void dart2JS(String input, String jsName) async {
-    var partPath = path.join(
-        Directory.current.path, input.replaceFirst('.bundle.json', '.dart'));
+  Future dart2JS(String input, String jsName) async {
+    var partPath = path.join(Directory.current.path, input.replaceFirst('.bundle.json', '.dart'));
     print('\u001b[33m [Fair Dart2JS] partPath => ${partPath} \u001b[0m');
     if (File(partPath).existsSync()) {
       try {
         var result = await dart2js.convertFile(partPath, true);
-        File(jsName)
-          ..writeAsStringSync(result);
+        File(jsName)..writeAsStringSync(result);
       } catch (e) {
         print('[Fair Dart2JS] e => ${e}');
       }
     }
   }
 
-  void compileBundle(PostProcessBuildStep buildStep, String bundleName) async {
+  Future compileBundle(PostProcessBuildStep buildStep, String bundleName) async {
     final bytes = await buildStep.readInputAsBytes();
     final file = File(bundleName)..writeAsBytesSync(bytes);
     if (file.lengthSync() > 0) {
@@ -70,8 +78,7 @@ class ArchiveBuilder extends PostProcessBuilder with FlatCompiler {
     }
     final buffer = StringBuffer();
     buffer.writeln('# Generated by Fair on ${DateTime.now()}.\n');
-    final source =
-        buildStep.inputId.path.replaceAll(inputExtensions.first, '.dart');
+    final source = buildStep.inputId.path.replaceAll(inputExtensions.first, '.dart');
     buffer.writeln('source: ${buildStep.inputId.package}|$source');
     final digest = md5.convert(bytes).toString();
     buffer.writeln('md5: $digest');
@@ -80,8 +87,7 @@ class ArchiveBuilder extends PostProcessBuilder with FlatCompiler {
       buffer.writeln('bin: ${buildStep.inputId.package}|${bin.data}');
     }
     buffer.writeln('date: ${DateTime.now()}');
-    File('${bundleName.replaceAll('.json', '.metadata')}')
-        .writeAsStringSync(buffer.toString());
+    File('${bundleName.replaceAll('.json', '.metadata')}').writeAsStringSync(buffer.toString());
 
     print('[Fair] New bundle generated => ${file.path}');
   }
@@ -93,12 +99,14 @@ class ArchiveBuilder extends PostProcessBuilder with FlatCompiler {
         continue;
       }
       if (entity.path.endsWith('.js') || entity.path.endsWith('.json')) {
-        final file = entity as File;
-        var filename = file.path.split('/').last;
+        final file = entity;
+        var filename = file.path.split(Platform.pathSeparator).last;
         final List<int> bytes = file.readAsBytesSync();
         archive.addFile(ArchiveFile(filename, bytes.length, bytes));
       }
     }
-    zipFile.writeAsBytesSync(ZipEncoder().encode(archive), flush: false);
+    var encode = ZipEncoder().encode(archive);
+    if (encode == null) return;
+    zipFile.writeAsBytesSync(encode, flush: false);
   }
 }
