@@ -27,11 +27,11 @@ mixin AppState {
 
   void setup(
     bool profile,
-    Map<String, FairDelegateBuilder> builder,
-    GeneratedModule generated,
-    Map<String, FairModuleBuilder> module,
+    Map<String, FairDelegateBuilder>? builder,
+    GeneratedModule? generated,
+    Map<String, FairModuleBuilder>? module,
   ) {
-    modules.addAll(module);
+    modules.addAll(module??{});
     GlobalState.instance().init(profile, builder);
     _proxy.addGeneratedBinding(generated);
   }
